Alert user that they are about to delete something
When user is about to delete something, an alert will be displaying asking if that's what they really want

+++ src/pages/albums/Index.js
@@ -3,6 +3,7 @@ import Navbar from "../../components/Navbar";
 import Footer from "../../components/Footer";
 import AlbumsTable from "../../components/AlbumsTable";
 import Popup from "../../components/Popup";
+import Swal from "sweetalert2";
 
 class AlbumsIndex extends React.Component {
   state = {
@@ -32,20 +33,39 @@ class AlbumsIndex extends React.Component {
     this.setState({ albums: data });
   }
 
-  async deleteAlbum(id) {
-    let formData = new FormData();
-    formData.append("id", id);
-    await fetch(`api/albumsApi/${id}`, {
-      method: "delete", 
-      body: formData,     
+  deleteAlbum(id) {
+    // album about to be deleted
+    const album = this.state.albums.find(album => album.id == id);
+
+    // sweetalert2 to alert user that he is about to delete an album
+    Swal.fire({
+      icon: 'warning',
+      title: 'Tem a certeza que pretende eliminar este álbum?',
+      text: `${album.title} - ${album.releaseYear}`,
+      footer: "ATENÇÃO: Esta ação é irreversível",
+      showCancelButton: true,
+      cancelButtonColor: "#9ca3af", // gray-400 from tailwindcss
+      confirmButtonColor: "#ef4444", // red-500 from tailwindcss
+      confirmButtonText: "Eliminar",
+      cancelButtonText: "Cancelar",
     }).then(async res => {
-      if(res.ok) {
-        await this.getAlbums();
-        return;
-      };
-      throw res.text();
-    }).catch(async err => {
-      this.setState({ fetchErr: true, fetchMsg: await err });
+      // user confirmed action
+      if (res.isConfirmed) {
+        let formData = new FormData();
+        formData.append("id", id);
+        await fetch(`api/albumsApi/${id}`, {
+          method: "delete", 
+          body: formData,     
+        }).then(async res => {
+          if(res.ok) {
+            await this.getAlbums();
+            return;
+          };
+          throw res.text();
+        }).catch(async err => {
+          this.setState({ fetchErr: true, fetchMsg: await err });
+        });
+      }
     });
   }
